refactor: create <CenterBlock /> component for code reusability
There are lots of places in the React application where it's needed to
show certain content in the center of the screen. Instead of writing
the same code over and over again, this commit introduces a component
called <CenterBlock />. It must accept props.children and optionally
props.className, which will be combined with the already specified
className.