feat(client): add a page to change a user's password
If a user wants to reset their password, they would follow this link:
http://(localhost)/forgot-password
They would enter their email, solve the captcha challenge and the
server would send them a link, by clicking on which the user would be
able to enter a new password and send it to the server. The server
would send a link that looks something like that:
http://localhost/auth/reset-password/random_reset_token12345

This commit implements this page that the user will follow.
When a user follows this link, instead of showing the reset password
form immediately, React will first make a request to the API to verify
that the token is valid. If it's not, an error will be presented to the
user suggesting to request another link to reset their password.
If the reset token is correct, a form is shown to the user (password +
confirm password + recaptcha). After sending the form the server will
change the user's password. If the reset token expired during filling
in the form, React will notify the user after submitting the form.

Why is this pleliminary step of verifying the reset token when the page
loads so important? Because if the user followed the link and the link
has actually expired, the user would find out about this only after
they fill in the form, solve the captcha challenge and send the form.
This would be very frustrating for users.

feat(client): add a page to change a user's password

If a user wants to reset their password, they would follow this link:
http://(localhost)/forgot-password
They would enter their email, solve the captcha challenge and the
server would send them a link, by clicking on which the user would be
able to enter a new password and send it to the server. The server
would send a link that looks something like that:
http://localhost/auth/reset-password/random_reset_token12345

This commit implements this page that the user will follow.
When a user follows this link, instead of showing the reset password
form immediately, React will first make a request to the API to verify
that the token is valid. If it's not, an error will be presented to the
user suggesting to request another link to reset their password.
If the reset token is correct, a form is shown to the user (password +
confirm password + recaptcha). After sending the form the server will
change the user's password. If the reset token expired during filling
in the form, React will notify the user after submitting the form.

Why is this pleliminary step of verifying the reset token when the page
loads so important? Because if the user followed the link and the link
has actually expired, the user would find out about this only after
they fill in the form, solve the captcha challenge and send the form.
This would be very frustrating for users.