feat: add PATCH /api/auth/change-password endpoint
This endpoint is used to reset a user's password.

It expected these fields in the request body:
 - recaptchaToken: string
 - resetToken: string // the token that the user will receive when they
request to send a link to their email
 - password: string // new password in plain text (the server will take
care of hashing it)

Possible responses:
    - Status code 422: something is wrong with the request body (some
field is incorrect or is missing, the resetToken is invalid etc)
    If the resetToken is invalid, the server response will contain
an array. One element of this array will look like this:
        {
            message: 'resetToken is either invalid or has expired'
            field: 'resetToken'
        }
    It will be useful for the user to display this error.
 - Status code 522: the client sent too many requests (the nginx web
server is configured to accept only 1 request per second on this route)
 - Status code 200: the password has been changed successfully.