feat: add GET /api/auth/is-reset-token-valid endpoint
This route is used to find out whether the reset token is valid. If the
token is valid, it can be used to reset the password of the account
that is tied to this token.

This route will be used for changing a user's password. When a user
requests a link to their email and then follow it, React.js will make a
request to the server to verify that the reset token is valid. If React
didn't do it, the user would have to fill in the form and then submit
it to find out that the token has expired. This would be very
frustrating for end users.

This route is not very useful for potential attackers, because even if
they somehow guess the reset token, although they would be able to
change the password, they still wouldn't know which account this token
is tied to, so it's useless for them to try to guess this token.

It accepts a reset token as a param, for example:
/api/auth/is-reset-token-valid/randomtoken12345

This route returns an object of this shape:
{
    isValid: boolean
}

Possible status codes:
 - 200 (the server always returns this code, regardless of whether
the reset token actually exists or not).
 - 503 (the client made too many requests to this route. The nginx
web server implements rate limiting for this route to prevent brute
force attacks. 1 request can be made per second).