feat: limit the request body size to 5 KB
By default, the 'body-parser' package is configured to accept body
requests up to 100 KB in length. This is way too much for this
application. As of now, this server only accepts JSON, and typically
its size is pretty small. There are no plans yet for the server to be
able to accept images, and even if it does start accepting them, it
will be done through AWS S3 or similar services in order to reduce the
load on the API server and thus to make the application more scalable.

If a request body is too large, a 413 status code will be returned.
The response body will look like this:
{
    errors: [{ message: 'Request body is too large' }]
}