feat: add client-side OAuth 2.0 flow with Google
When a user navigates to /auth/sign-in or /auth/sign-up, they will be
presented with 2 buttons: authenticate via email or via Google.
It doesn't matter if the user wants to sign up or sign in, because if
they want to do so with Google, they will be redirected to the same
link: /auth/log-in-with/google

By following this link, the user will be redirected to the Google
authorization server, and after consenting the user will be redirected
to /auth/oauth-callback. The React application will send the request to
the server to the POST /api/auth/oauth-callback route, and the server
will respond with a refresh token and an access token.