fix(<TopHeader />): listen to auth with useSelector instead of props
Previously, the <TopHeader /> component was accepting the
'isAuthenticated' prop to determine whether the user is authenticated
or not. This was done to minimize the number of subscriptions to the
Redux store. However, this will lead to bugs when implementing the
OAuth 2.0 flow in the future.
This commit fixes this bug by subscribing to the Redux store.