refactor: split models into new submodels
This commit splits models into smaller ones. This was done to reduce
the size of different models and to improve modularity of the code, as
well as to make the models more maintainable.

Logic related to sending emails was put in the "services" folder. Now
any email logic is handled in the services/email.service.ts file.

Controllers were also updated to reflect these changes.