feat!: add POST /api/auth/logout endpoint
Add a new /logout route to the API that clears the refreshToken cookie,
deletes it from the database and sends a 204 response. Also, change the
path of the refreshToken cookie from /api/auth/refresh to /api/auth to
allow it to be sent to the /logout route.

BREAKING CHANGE: The path of the refreshToken cookie has been changed
from /api/auth/refresh to /api/auth.