style: move the "not activated" signin message to the bottom of the center block
When a user attempts to sign in with an inactive account, the error
message is now displayed below the "sign in" button along with a button
to resend the activation link. This change improves the visual
hierarchy of the sign-in form and highlights the action the user needs
to take to activate their account.

This change doesn't affect the functionality of the sign-in form.