feat: add "Explore Our Products" block
Added a block that shows featured products on the landing page.
If an individual product has run out, the product block will be
semi-visible. If the product is running out, it will be shown as well.

GraphQL is used to fetch the products. To make the process of working
with GraphQL in React type safe, I used graphql-codegen to generate
.ts files. These files will be in the client/src/__generated__ folder.
One important caveat is that graphql-codegen requires our REST API to
work. This is why I created a bash script that waits for the API to
start up before starting the React.js development server.