feat: add isAdmin status to the Redux store
This commit adds a new property to the 'auth' slice called 'isAdmin'.
It designates whether a user is an admin or not. This status is
retrieved by decoding the JWT accessToken. This 'isAdmin' property will
be used in the future to prevent users from accessing protected routes.

There is no need to worry about verifying this token, because we got it
from the server, so we can safely assume that it is valid. Even if for
some reason the token is malformed, the user won't be able to perform
any serious action, because the API routes are protected.