refactor(<SubmitButton />): add a new prop for considering form dirtiness
The submit button component has been refactored to introduce a new prop
called "considerDirtyProp". By default, this prop is set to true,
indicating that the button should be disabled if the form is not dirty,
i.e., if the user hasn't made any changes to the form.
If "considerDirtyProp" is set to false, the button will not consider
the form dirtiness when deciding whether to disable the button, but it
will still consider other factors such as whether the form is being
submitted (isSubmitting), if the loading state is active (isLoading
parameter), or if there are any form validation errors.

This change allows for more flexibility in using the submit button
component based on the specific requirements of the form.