feat(API): allow fetching product category in GraphQL queries
This commit enables the retrieval of the product category in GraphQL
queries. Previously, the product category information was not included
in the queries, but now it is allowed for both admin queries and basic
queries, accessible to all users.

By updating the GraphQL schema and associated queries, the product
category data can now be fetched alongside other product details. This
enhancement provides more comprehensive information and flexibility
when working with product-related data in the application.