feat(API): add an endpoint to fetch product categories
GET /api/product-categories
Returns all product categories.
Response: { categories: string[] }
Who can access this endpoint: anyone

The motivation for this endpoint is to facilitate category selection
when editing products in the admin interface. By fetching the
categories, the React app can present them as options in the <select>
tag, allowing admins to easily change the category of a product.
Additionally, this endpoint lays the groundwork for future
enhancements, such as displaying categories on the main page and
enabling filtering of products by category.