refactor: simplify validation schema passing in FormInput using context
This commit refactors the FormInput component to simplify the way
validation schemas are passed. Previously, a validation schema had to
be provided individually to each FormInput component, which could
become repetitive and cumbersome, especially when dealing with forms
containing multiple inputs.

To address this, a SchemaContext is introduced, leveraging the Context
API. The SchemaContext acts as a provider, allowing a validation schema
to be provided once and shared across multiple FormInput components.
Now, instead of explicitly passing the schema to each FormInput, it can
be set by wrapping the form like this:
<SchemaContext.Provider value={validationSchema}>
    ...
</SchemaContext.Provider>

This refactor improves code organization, reduces redundancy, and
enhances the reusability of the FormInput component. It provides a
cleaner and more efficient approach to managing validation within
Formik forms.