refactor(<FormInput />): optimize handleChange and handleBlur functions
This commit refactors the <FormInput /> component to optimize the
"handleChange" and "handleBlur" functions. Previously, these functions
contained similar code, leading to redundancy. To eliminate
duplication, the common functionality has been encapsulated in two new
functions: "getInputData" and "validateSeparateInput".

The "getInputData" function retrieves the input name and value from the
input element. If the input type is "file", the value is now stored as
an object with the structure { size: number; type: string }. This
change allows for validating the type and size of an image using Yup
while still using the <FormInput /> component.

The "validateSeparateInput" function handles the separate validation of
an input, taking into account special cases like validating the
"password" and "confirmPassword" fields together.

With these optimizations, the "handleChange" and "handleBlur" functions
have become leaner and more efficient, reducing code duplication and
improving the maintainability of the component.

No functional changes were made, and this component can still be used
with any other input types. The commit solely focuses on improving the
internal implementation of <FormInput />.