feat(<FormInput />): add support for 'min', 'max', and 'step' attributes
This commit enhances the functionality of the FormInput component to
support the usage of 'min', 'max', and 'step' attributes for inputs
with the type "number". Previously, these attributes were not
available, but now it's possible to utilize them to specify the minimum
value, maximum value, and incremental step for number inputs. This
improvement extends the flexibility and customization options for
handling numeric input fields within the FormInput component.