feat(client): implement admin add-product form
This commit adds a new page (/add-product) that allows admins to enter
data for a new product. This form has the following fields:
- title of the product (<= 200 characters)
- price of the product (> 0)
- initial image (.png, <= 500 KB)
- additional image (.png, <= 500 KB)
- quantity in stock (>= 0)
- short description (<= 300 characters)
- category (not validated because the list of allowed categories is
fetched from the server. Besides, there is little sense in validating
this field because simple users won't be able to access this form
anyway. Sending API requests directly is also not the option for
malicious users, because the API server validates who makes a request).

If the following conditions are met and the user is trying to submit
the form, the following will happen:
- React will try to get 2 presigned URLs for image upload (one for the
initial image, one for the additional image). This is done by making a
request to the API and putting the image name as part of the request
query. React will generate random image names for these images. There
is no harm in generating the image names on the client side, because
adding products is only allowed for admins, and if someone becomes an
admin without authorization, we'll have much bigger problems than image
name generation.
- The initial image and the additional image are then uploaded to AWS S3
via these presigned URLs.
- Once the images are uploaded, the main GraphQL addProduct request is
made. All the information about the product including image names will
be included in the request.
- If everything goes as expected, a success block will be shown.
Otherwise, an error block will be shown.