feat(<FormInput />): add forwardRef support to access input
This commit adds the ability to access the underlying input element in
the <FormInput /> component via forwardRef. This is useful as it adds
more flexibility in handling forms. If you need to get access to any
input that's inside the FormInput component, it's now possible to use
the useRef hook with this component.