refactor(API addProduct mutation): accept image names instead of URLs
In this commit, I have refactored the "addProduct" mutation in the
GraphQL schema to improve the efficiency of image uploads in the
e-commerce application. Previously, the mutation required the admin to
provide the URLs of the initial and additional images for a product.
However, this approach proved to be inefficient and required additional
API requests to retrieve the full image URLs.

To address this, I have modified the "addProduct" mutation to now
accept "initialImageName" and "additionalImageName" as parameters
instead of URLs. These parameters represent the filenames of the images
being uploaded to S3 (e.g., "someimage.png").

With this change, the flow of uploading images has been streamlined.
The admin now specifies the simple fields such as "title," "price,"
etc. by using React, and uploads the images on the client side. The
main request to the "addProduct" mutation is put on hold, allowing
React to fetch the presigned URLs for uploading the initial and the
additional images. React then uploads the images to AWS S3 using these
presigned URLs.

To retrieve the full URLs of the uploaded images, the API server will
now deduce the image URLs based on the provided image names and adds
the product to the database in a single step. This eliminates the need
for additional API requests to fetch image URLs.

By transitioning from specifying image URLs to image names in the
"addProduct" mutation, we have improved the efficiency and simplified
the process of uploading images in the e-commerce application.

Important note: when you specify "initialImageName" and
"additionalImageName" in the "addProduct" mutation, these images must
already be uploaded to S3, otherwise the product will have invalid
images pointing to nothing.