fix(modal window): add 'position: fixed'
Previously, the modal window and the backdrop associated with it had
'position: absolute'. However, if a page was so large that a scrollbar
appeared there, the modal window wouldn't be visible unless you
scrolled all the way to the top of the page.

This commit fixes this behaviour by changing the 'position' property to
'fixed'.