feat!(<Modal />): enable forwardRef for accessing the underlying div
This commit introduces support for accessing the underlying div element
within the <Modal /> component using forwardRef.

This capability is essential for utilizing the 'body-scroll-lock' npm
package, which necessitates specifying the element for persistent
scrolling, such as within a modal, lightbox, flyout, or navigation.

Note: Enabling forwardRef for <Modal /> also required modifying the
<Card /> component, as it relies on forwardRef internally. As a result,
obtaining the underlying HTML element of the <Card /> component is now
possible as well.