fix(<AddProductForm />): set initial value for category field to prevent foreign key constraint error
When an admin attempts to add a new product in the e-commerce
application, an error message was displayed:
"insert or update on table 'products' violates foreign key constraint
'fk_products_product_categories'".
The issue stemmed from the form used for adding products, which
included a <select> element to choose a category.

The bug occurred because although the <select> element had a default
option, formik incorrectly thought that the category had not been
chosen. Consequently, when the form was submitted, an empty string was
sent as the value for the "category" field. PostgreSQL then returned an
error, stating that there was no category with an empty string as its
identifier.

To resolve this issue, the bug was fixed by ensuring that the initial
value of the "category" field in the form is set to the default option
of the <select> element. This ensures that a valid category is always
submitted, preventing the foreign key constraint error.